requirements.txt
flask==2.3.3
pandas>=2.0.0
numpy==1.24.3
matplotlib==3.7.2
python-binance>=1.0.19
python-dotenv==1.0.0
mplfinance>=0.12.9b7